more work on #259383: do not actually remove the apache configuration file when disabling a site.
just removing the config file will not work since the default vhost kicks in and still allows the disabled site to work. instead, we
create a custom vhost configuration file that redirects to a "disabled" url (that we should make configurable eventually).

right now that url is just $base_url/disabled, which gives a page not found, but i'll work more on that