Revert "initialize mysql credentials from drush bootstrap so that we can work on non-verified sites easily"
Provision is managing the master database connection here, drush's db credentials are the details for the SITE, and will only be available at drush_bootstrap_drupal_site.

we need the master db connection from drush_bootstrap_drush already, and we will never ever ever be able to make use of drush' code in this regard because what we are doing is completely outside of the Drupal context.

This reverts commit 5439e6d01d32fd7351c404a6bf74c11a028ee8a4.