#683800 - fix removal of symlink on site removal
this was failing because provision_path_unlink was checking the file existence (file_exists) before removing the symlink. in this case, the file really *didn't* exist because the link target was already removed.

instead of checking, we just try to delete the file, if it fails, it will return false anyways.

we do the same with rmdir while we're here.

also make sure we work on the trimmed file path