Introduce magic getter/setters into the service, along with the owner property.
Constructing a service requires an owner alias name, which is at the moment only
@self, or the left hand alias passed in a drush command.

If properties aren't defined in the service class, when you get or set them,
they will be set on the owner's environment object.

In the backend I am using d($this->owner) to access the properties.