do not attempt to chown files/ dir during verification
the rationale here is that we're not running as root, and therefore
chown is likely to fail, unless the file already has the right owner,
in which case a chown is not necessary.

To quote chown(2):

       Only a privileged process (Linux: one with the CAP_CHOWN
       capability) may change the owner of a file.  The owner of a
       file may change the group of the file to any group of which
       that owner is a member.  A privileged process (Linux: with
       CAP_CHOWN) may change the group arbitrarily.

So this was just a cause of trouble without any possible benefit
unless we run with capabilities, and we shouldn't assume this (right
now?).