1081536 rollback platform apache configs
when drush make fails, the apache config is already created and wasn't
cleaned up

we hook into the rollback by adding a new magic hook through
command_invoke() that will cascade through the current context and call
the right rollback hook.

we implement one for the platform, but we probably need to implement one
for servers and sites too.