Revert "1081536 rollback platform apache configs"
This reverts commit f7a071135d626bcef4879dc7e8687fd888d8383b.

The problem with this approach is that if we have sites in the platform
and fail to verify it, removing the apache config would actually break
the platform even more. We do not want that.

What we want is a provision-install on platforms that can be rolled back
properly too. It's harder than this.