make this a native package
while Aegir is desgined to run on other systems than Debian, it is
currently *not* uploaded into Debian itself, and therefore there is
only one source of packages, debian.aegirproject.org.

this will mean we will need to make a new release every time we want to publish changes to the Debian package, but this annoyance outweights the trouble of having to maintain 4 different branches (1.x, 2.x and debian, debian2x) to have a 2.x debian package.

once we start looking again at sending the Debian packages directly into the main Debian archive, we can re-create the debian branch and make non-native packages there. for example, 2.x could eventually have no debian/ directory and the debian branch would be solely for the 2.x release...

in the meantime the debian branch is not to be used.