remove IP based certificate management
IP addresses are managed through the frontend now and passed down
here.

we are not properly handling IP addresses however, as there is a
duplication between the site_ip_addresses and ip_address field.

this deprecates the assign_certificate_ip(), get_ip_certificate() and
free_certificate_ip() in the Provision_Service_http_ssl class. the
clear_certs() function in Provision_Service_http_ssl is also
deprecated.

we also refactor the way the uri-level receipt files are managed to
encapsulate the code in functions and automatically cleanup
certificates within that allocation code.