revert last changes to release process
we are keeping the old method of installing from git, which means we
will install post-alpha2 code when there are commits to the frontend,
but meh - we currently can't install alpha2 at all from git, so that's
better.

we at least need to download the hostmaster profile in aegir.make, but
then that means it will recurse into hostmaster.make, so it means we
need to fiddle with *that* code if we want have sticky branches there.

the other alternative is to build from the distribution, but that's
not supported by Drush make, only by pm-download.