properly manage multiple configuration files per context
it seems this was just incorrectly done: the _config entry wasn't used
as an array while $configs is an array of arrays.

since _config is private I feel safe in changing the way it is
structured and used internally, and externally the API doesn't really
change, it just works. :)

this is to fix the regression introduced in e918bcc, see #2000038