Fix for backup_migrate support.
Otherwise web server can't write there, so even if we use Drush for cron, and it will run as an Aegir backend user, which has the write access there (and also limited shell user can write there) the module itself will refuse to work on scheduled cron run, because the backup directories are expected to be writable by the web server. This is required also when standard wget method for cron is used.