Issue #2098389: Record enabled Hosting Features in /var/aegir/.drush/drushrc.php.
We want to be able to block unwanted Drush hooks from firing, especially those in
extensions where the front-end module is not enabled. To do so, we need to know
which Hosting Features are enabled very early in the Drush bootstrap process, since
blocking the unwanted hooks from firing is accomplished with hook_drush_load()'s.