move subdir implementation to the Apache service
the reasoning is that we need to operate on multiple servers, which
can't be done outside of services.

furthermore, we had to move to a Config class for sync to work, and we
hack up that class badly so that it is forced to generate multiple
configuration files instead of just one, which is impure and weird,
but seems to work in my first tests.

we still need to reimplement sites.php and disabled sites support, but
multi-server should work better now

see #2020075 and #2020079