don't go around destroying apache configs on upgrades or removal
the rationale for upgrades is obvious: even if the next package will
restore the link, the upgrade may fail and then all sites go
down. even if it doesn't fail, sites go down for the duration of the
upgrade: bad. this affects 1.x to 2.x upgrades.

the rationale for remove is that we don't completely destroy the data,
so the sites are expected to remain.

we also cleanup after ourselves in special corner cases
(abort-install, disappear) as well.