don't allow d() to run before drush is fully initialized
we add safety check so that we don't get caught running d() before
drush is fully initialized. otherwise, the proper services hooks may
not be properly declared and found by drush, so the object cache
($instances) would be initialised with incomplete data.

it is possible that more initialisation we need happens after
DRUSH_BOOTSTRAP_NONE, but this is already better than nothing.

it took me and ergonlogic three freaking full days of headbanging on
keyboards to figure that stuff out.